Extracted some LiteralTransformer tests to StringArrayStorage.spec.ts
