+++ test/functional-tests/storages/string-array-storage/fixtures/one-string.js
@@ -1,0 +1,1 @@
+var test = 'test';
