+++ test/functional-tests/storages/string-array-storage/fixtures/three-strings.js
